adds finance api to sdk

+++ lib/xero-ruby/models/accounting/budget_balance.rb
@@ -42,8 +42,8 @@ module XeroRuby::Accounting
     def self.openapi_types
       {
         :'period' => :'Date',
-        :'amount' => :'Integer',
-        :'unit_amount' => :'Integer',
+        :'amount' => :'BigDecimal',
+        :'unit_amount' => :'BigDecimal',
         :'notes' => :'String'
       }
     end
